Free shipping cart rule order-carrier
I created a website for a client and he created a free shipping cart rule with restrictions on France and Belgium. Unfortunately, my test account had four delivery addresses (One in France, Belgium, and Germany).

You made a call in Cart.php in the function getDeliveryOptionList() to reach for cart rules with free shipping ($cart_rules = CartRule::getCustomerCartRules(Context::getContext()->cookie->id_lang, Context::getContext()->cookie->id_customer, true, true, false, $this);).

Unfortunately in this function, you make a foreach countries (for all addresses) and if you have an adress forbidden (even if it's not the delivery address) it unsets the cart rule. This patch reach for id_address_delivery if it exists and get cart rules free shipping only for it.